pipeline/requirements.txt
# pipeline/requirements.txt
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
schedule==1.2.0
pyspark==3.5.2
